backend/requirements.txt
Django==4.2.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-pandas==0.6.7
djangorestframework==3.14.0
keras==2.14.0
numpy==1.26.4
pandas==2.0.3